@@ -2,7 +2,7 @@ name: flipper_rw
 description: "A new Flutter project."
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 1.170.4252223233965+1747585288
+version: 1.170.4252223233968+1748192469
 
 
 # 
@@ -20,7 +20,7 @@ dependencies:
   
   firebase_crashlytics: ^4.3.4
   flutter_local_notifications: ^19.1.0
-  intl: ^0.19.0
+  intl: ^0.20.2
   injectable: ^2.5.0
   get_it: ^7.7.0
   stacked: ^3.4.3
@@ -166,7 +166,7 @@ dependencies:
   amplify_storage_s3: ^2.6.1
   shared_preferences: ^2.5.2
   easy_localization: ^3.0.7 
-  supabase_flutter: ^2.8.4
+  supabase_flutter: ^2.9.0
   sqflite_common_ffi: ^2.3.4+4
   # https://github.com/alextekartik/flutter_app_example/blob/master/demo_sqflite/tool/windows/windows_release_info.md
   sqflite_common_ffi_web: ^1.0.0
@@ -224,7 +224,7 @@ msix_config:
   publisher_display_name: yegobox ltd
   identity_name: yegobox.yegoboxflipper
 
-  msix_version: 1.170.4115.0
+  msix_version: 1.170.4118.0
   publisher: CN=B0F334A1E-8AET-4B2C-BD1B-F239D4F791AE
   logo_path: "./assets/icon.png"
   start_menu_icon_path: "/assets/icon.png"
