@@ -12,7 +12,7 @@ dependencies:
   meta: ^1.3.0
   universal_platform: ^1.0.0+1
   stacked_services: ^1.6.0
-  firebase_auth: ^5.5.2
+  firebase_auth: ^5.5.3
   stacked: ^3.4.2
   built_value: ^8.1.4
   azlistview: ^2.0.0
@@ -42,7 +42,7 @@ dependencies:
   talker_flutter: ^4.1.5
   rxdart: ^0.28.0
   collection: ^1.19.0
-  intl: ^0.19.0
+  intl: ^0.20.2
   amplify_core: ^2.6.1
   amplify_api: ^2.6.1
   amplify_flutter: ^2.6.1
@@ -55,15 +55,15 @@ dependencies:
   sqlite3_flutter_libs: ^0.5.31
   permission_handler: ^11.3.1
   json_annotation: ^4.9.0
-  supabase_flutter: ^2.8.4
+  supabase_flutter: ^2.9.0
   supabase_models: any
-  
   printing: ^5.14.2
   pdf: ^3.11.3
   injectfy: ^1.0.1
   mock_supabase_http_client: 0.0.3+1
   sqlite3_web: ^0.3.0
   synchronized: ^3.3.0+3
+  email_validator: ^2.1.17
   
 
 dependency_overrides:
