@@ -16,7 +16,7 @@ dependencies:
   share_plus: ^10.1.4
   firebase_messaging: ^15.2.2
   image: ^4.0.17
-  firebase_auth: ^5.5.2
+  firebase_auth: ^5.5.3
   geolocator: ^13.0.2
   geocode: ^1.0.3
   firebase_analytics: ^11.4.4
@@ -24,7 +24,7 @@ dependencies:
   location: ^8.0.0
   google_sign_in: ^6.2.2
   http: ^1.2.2
-  intl: ^0.19.0
+  intl: ^0.20.2
   path_provider: ^2.1.5
   path: ^1.8.2
   googleapis: ^11.2.0
@@ -51,7 +51,7 @@ dependencies:
   launcher_entry: ^0.1.1
   internet_connection_checker: ^1.0.0+1
   talker_flutter: ^4.1.5
-  supabase_flutter: ^2.8.4
+  supabase_flutter: ^2.9.0
   firebase_core: ^3.13.0
   # cbl_flutter_ce: ^3.1.3
   # cbl_flutter: ^3.1.3
