@@ -12,7 +12,7 @@ dependencies:
   brick_offline_first_with_supabase: ^2.0.0
   brick_sqlite: ^4.0.1
   brick_supabase: ^2.0.0
-  supabase_flutter: ^2.8.4
+  supabase_flutter: ^2.9.0
   sqflite_common: ^2.5.4+6 
   sqflite_common_ffi: ^2.3.4+4
   path_provider: ^2.1.5
