@@ -102,6 +102,41 @@ class AuthState {
   }
 }
 
+/// Authentication state for reactive UI updates
+class AuthenticationState {
+  final bool isAuthenticating;
+  final String? errorMessage;
+  final bool isComplete;
+
+  AuthenticationState({
+    this.isAuthenticating = false,
+    this.errorMessage,
+    this.isComplete = false,
+  });
+
+  factory AuthenticationState.idle() {
+    return AuthenticationState(isAuthenticating: false);
+  }
+
+  factory AuthenticationState.authenticating() {
+    return AuthenticationState(isAuthenticating: true);
+  }
+
+  factory AuthenticationState.error(String message) {
+    return AuthenticationState(
+      isAuthenticating: false,
+      errorMessage: message,
+    );
+  }
+
+  factory AuthenticationState.complete() {
+    return AuthenticationState(
+      isAuthenticating: false,
+      isComplete: true,
+    );
+  }
+}
+
 class AuthController {
   final _authStateController = StreamController<AuthState>.broadcast();
 
@@ -145,11 +180,19 @@ class _AuthState extends State<Auth> {
   String? _errorMessage;
   late LoginViewModel _loginViewModel;
 
+  // Authentication state controller for reactive UI updates
+  final _authenticationStateController =
+      StreamController<AuthenticationState>.broadcast();
+
+  // Dialog management
+  bool _isAuthDialogShowing = false;
+
   @override
   void initState() {
     super.initState();
     _loginViewModel = LoginViewModel();
-    
+
+    // Listen to the auth controller state changes
     _authController.authState.listen((state) {
       setState(() {
         _isLoading = state.status == AuthStatus.loading;
@@ -157,36 +200,119 @@ class _AuthState extends State<Auth> {
       });
 
       if (state.status == AuthStatus.success) {
-        showDialog(
-          context: context,
-          barrierDismissible: false, // Prevent closing by tapping outside
-          builder: (BuildContext context) {
-            return const LoadingDialog(message: 'Finalizing authentication...');
-          },
-        );
-        
         // Set up Firebase auth state listener to handle the complete login process
         _handleAuthStateChanges();
       }
     });
+
+    // Listen to authentication state for showing/hiding loading dialog
+    _authenticationStateController.stream.listen((state) {
+      if (state.isAuthenticating && !_isAuthDialogShowing) {
+        _showAuthenticationDialog();
+      } else if (!state.isAuthenticating && _isAuthDialogShowing) {
+        _hideAuthenticationDialog();
+      }
+
+      if (state.errorMessage != null) {
+        _authController.notifyError(state.errorMessage!);
+      }
+    });
+  }
+
+  // Show authentication loading dialog
+  void _showAuthenticationDialog() {
+    _isAuthDialogShowing = true;
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (BuildContext context) {
+        return const LoadingDialog(message: 'Finalizing authentication...');
+      },
+    ).then((_) {
+      _isAuthDialogShowing = false;
+    });
+  }
+
+  // Hide authentication loading dialog
+  void _hideAuthenticationDialog() {
+    if (_isAuthDialogShowing && mounted) {
+      Navigator.of(context, rootNavigator: true).pop();
+      _isAuthDialogShowing = false;
+    }
   }
 
   /// Handles user authentication state changes and login flow
   Future<void> _handleAuthStateChanges() async {
-    FirebaseAuth.instance.authStateChanges().listen((user) async {
-      if (user == null) return;
-      
-      final bool pinLoginEnabled = (await ProxyService.box.pinLogin()) ?? false;
-      final bool authIsComplete = (await ProxyService.box.authComplete());
-      final bool shouldProcessLogin = !pinLoginEnabled && !authIsComplete;
+    // Notify UI that authentication is in progress
+    _authenticationStateController.add(AuthenticationState.authenticating());
 
-      if (!shouldProcessLogin) return;
+    FirebaseAuth.instance.authStateChanges().listen((user) async {
+      if (user == null) {
+        _authenticationStateController.add(AuthenticationState.idle());
+        return;
+      }
 
       try {
-        final loginData = await _loginViewModel.processUserLogin(user);
+        // Only check if authentication is already complete
+        // PIN login should not prevent email/phone login
+        final bool authIsComplete = (await ProxyService.box.authComplete());
+
+        // Log authentication state for debugging
+        talker.info(
+            'Auth state: authIsComplete=$authIsComplete, user=${user.uid}');
+
+        if (authIsComplete) {
+          talker.info('Skipping login process: user is already authenticated');
+
+          // Update UI state to idle (will automatically dismiss loading UI)
+          _authenticationStateController.add(AuthenticationState.idle());
+
+          // Navigate to startup view since auth is already complete
+          _routerService.clearStackAndShow(StartUpViewRoute());
+          return;
+        }
+
+        // Set a timeout to prevent indefinite hanging
+        bool timeoutOccurred = false;
+        Future.delayed(Duration(seconds: 30)).then((_) {
+          if (_isAuthDialogShowing) {
+            timeoutOccurred = true;
+            talker.warning('Authentication process timed out after 30 seconds');
+            _authenticationStateController.add(AuthenticationState.error(
+                'Authentication timed out. Please try again.'));
+          }
+        });
+
+        // Process login with retry for network issues
+        int retryCount = 0;
+        Map<String, dynamic>? loginData;
+
+        while (retryCount < 2 && !timeoutOccurred) {
+          try {
+            loginData = await _loginViewModel.processUserLogin(user);
+            break; // Success, exit retry loop
+          } catch (e) {
+            if (e.toString().contains('network') && retryCount < 1) {
+              // Only retry once for network errors
+              retryCount++;
+              talker.info(
+                  'Network error during login, retrying (${retryCount}/2)...');
+              await Future.delayed(Duration(seconds: 2)); // Wait before retry
+            } else {
+              rethrow; // Not a network error or max retries reached
+            }
+          }
+        }
+
+        if (loginData == null) {
+          throw Exception('Failed to process login after retries');
+        }
+
         final Pin userPin = loginData['pin'];
         final IUser userData = loginData['user'];
-        await _loginViewModel.completeLoginProcess(userPin, _loginViewModel, user: userData);
+
+        await _loginViewModel.completeLoginProcess(userPin, _loginViewModel,
+            user: userData);
 
         // Track login event with PosthogService
         PosthogService.instance.capture('login_success', properties: {
@@ -194,13 +320,14 @@ class _AuthState extends State<Auth> {
           'user_id': user.uid,
           'email': user.email ?? user.phoneNumber!,
         });
-        
-        // Note: The loading dialog will be closed automatically during navigation
+
+        // Update UI state to idle (will automatically dismiss loading UI)
+        _authenticationStateController.add(AuthenticationState.idle());
       } catch (e, s) {
-        // Ensure the dialog is closed on error
-        if (mounted) {
-          Navigator.of(context, rootNavigator: true).pop();
-        }
+        talker.error('Authentication error: $e');
+        // Update UI with error state (will automatically show error and dismiss loading)
+        _authenticationStateController
+            .add(AuthenticationState.error(e.toString()));
         _loginViewModel.handleLoginError(e, s);
       }
     });
@@ -208,6 +335,7 @@ class _AuthState extends State<Auth> {
 
   @override
   void dispose() {
+    _authenticationStateController.close();
     _authController.dispose();
     super.dispose();
   }
@@ -251,19 +379,66 @@ class _AuthState extends State<Auth> {
   Future<void> _handleMicrosoftLogin() async {
     _authController.notifyLoading();
     try {
+      // Configure Microsoft provider to allow any Microsoft account
       final provider = MicrosoftAuthProvider();
-      provider.addScope('mail.read');
 
-      final userCredential =
+      // Set 'common' tenant to allow any Microsoft account (personal or organizational)
+      provider.setCustomParameters({
+        'tenant': 'common',
+        'prompt':
+            'select_account', // Ensures user can select from multiple accounts
+      });
+
+      // Add necessary scopes for profile access
+      provider.addScope('user.read');
+      provider.addScope('openid');
+      provider.addScope('profile');
+      provider.addScope('email');
+
+      // Log the authentication attempt
+      talker.info(
+          'Starting Microsoft OAuth sign-in process with multi-tenant configuration');
+
+      // Perform the authentication
+      final authCredential =
           await FirebaseAuth.instance.signInWithProvider(provider);
-      if (userCredential.user != null) {
+      talker.info('Microsoft credential obtained successfully');
+
+      if (authCredential.user != null) {
+        talker
+            .info('Microsoft sign-in successful: ${authCredential.user?.uid}');
         _authController.notifySignedIn();
       } else {
         _authController.notifyError("Sign in failed");
       }
-    } catch (e) {
+    } on FirebaseAuthException catch (e) {
+      // Handle specific Firebase Auth errors
+      talker.error(
+          'Microsoft login FirebaseAuthException: ${e.code} - ${e.message}');
       Sentry.captureException(e, stackTrace: StackTrace.current);
-      _authController.notifyError("Microsoft login failed");
+
+      // Handle user cancellation gracefully
+      if (e.code == 'popup-closed-by-user' ||
+          e.code == 'cancelled' ||
+          e.code == 'web-context-canceled') {
+        _authController.notifySignedOut();
+      } else if (e.code == 'unauthorized-domain') {
+        _authController.notifyError(
+            "Authentication domain not authorized. Please contact support.");
+      } else if (e.code == 'user-disabled') {
+        _authController.notifyError("This account has been disabled.");
+      } else if (e.code == 'account-exists-with-different-credential') {
+        _authController.notifyError(
+            "An account already exists with the same email address but different sign-in credentials.");
+      } else {
+        _authController.notifyError("Microsoft login failed: ${e.message}");
+      }
+    } catch (e) {
+      // Log detailed error information for other exceptions
+      talker.error('Microsoft login error: ${e.toString()}');
+      Sentry.captureException(e, stackTrace: StackTrace.current);
+      _authController
+          .notifyError("Microsoft login failed. Please try again later.");
     }
   }
 
