@@ -2,12 +2,14 @@ import 'dart:developer';
 import 'dart:convert';
 import 'package:flipper_models/helperModels/business_type.dart';
 import 'package:flipper_models/db_model_export.dart';
+import 'package:flipper_routing/app.locator.dart';
+import 'package:flipper_routing/app.router.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:stacked/stacked.dart';
 import 'package:overlay_support/overlay_support.dart';
-
+import 'package:stacked_services/stacked_services.dart';
 import 'package:flutter_form_bloc/flutter_form_bloc.dart';
 import 'package:flipper_services/proxy.dart';
 
@@ -105,8 +107,6 @@ class AsyncFieldValidationFormBloc extends FormBloc<String, String> {
           tinNumber.value.isEmpty ? "999909695" : tinNumber.value;
       signupViewModel.businessType = businessTypes.value!;
       await signupViewModel.signup();
-
-      // _routerService.navigateTo(StartUpViewRoute(invokeLogin: true));
       emitSuccess();
     } catch (e) {
       signupViewModel.stopRegistering();
