@@ -2,9 +2,7 @@ import 'dart:convert';
 import 'dart:developer';
 import 'package:flipper_models/rw_tax.dart';
 import 'package:flipper_models/secrets.dart';
-import 'package:flipper_services/proxy.dart';
 import 'package:http/http.dart' as http;
-import 'package:sentry/sentry.dart';
 
 extension RWTaxEmailExtension on RWTax {
   Future<void> sendEmailNotification(
@@ -24,28 +22,20 @@ extension RWTaxEmailExtension on RWTax {
 
     request.headers.addAll(headers);
 
-    await request.send();
+    ///TODO: resum the feature in the future
+    /// await request.send();
   }
 
   Future<bool> sendRequestAndHandleResponse(http.Request request) async {
     try {
       http.StreamedResponse response = await request.send();
-      final requestData = request.body.toString();
+      // final requestData = request.body.toString();
       final responseData = await response.stream.bytesToString();
 
       if (response.statusCode == 200) {
         log(responseData);
 
-        final sentryId = await Sentry.captureMessage("EBM-JSON");
-
-        final userFeedback = SentryUserFeedback(
-          eventId: sentryId,
-          comments: requestData,
-          email: ProxyService.box.getUserPhone(),
-          name: ProxyService.box.getUserPhone(),
-        );
-
-        Sentry.captureUserFeedback(userFeedback);
+        // final sentryId = await Sentry.captureMessage("EBM-JSON");
 
         // await sendEmailNotification(
         //     requestData: requestData, responseData: responseData);
