@@ -35,7 +35,8 @@ mixin CapellaBusinessMixin implements BusinessInterface {
   }
 
   @override
-  Future<List<Business>> businesses({required int userId}) async {
+  Future<List<Business>> businesses(
+      {required int userId, bool fetchOnline = false}) async {
     throw UnimplementedError('businesses needs to be implemented for Capella');
   }
 
