@@ -146,7 +146,8 @@ mixin AuthMixin implements AuthInterface {
 
   // Required methods that should be provided by other mixins
   @override
-  Future<List<Business>> businesses({required int userId});
+  Future<List<Business>> businesses(
+      {required int userId, bool fetchOnline = false});
 
   @override
   Future<bool> firebaseLogin({String? token}) async {
@@ -226,7 +227,7 @@ mixin AuthMixin implements AuthInterface {
 
     // Avoid unnecessary sync if payment is already marked as complete
     if (!isPaymentCompletedLocally) {
-      final isPaymentComplete = await ProxyService.realmHttp.isPaymentComplete(
+      final isPaymentComplete = await ProxyService.httpApi.isPaymentComplete(
         flipperHttpClient: flipperHttpClient,
         businessId: businessId,
       );
@@ -249,7 +250,7 @@ mixin AuthMixin implements AuthInterface {
 
   Future<IUser> _authenticateUser(
       String phoneNumber, Pin pin, HttpClientInterface flipperHttpClient,
-      {bool forceOffline = false}) async {
+      {bool forceOffline = false, bool freshUser = false}) async {
     List<Business> businessesE = await businesses(userId: pin.userId!);
     List<Branch> branchesE = await branches(businessId: pin.businessId!);
 
@@ -303,10 +304,19 @@ mixin AuthMixin implements AuthInterface {
         await sendLoginRequest(phoneNumber, flipperHttpClient, apihub);
 
     if (response.statusCode == 200 && response.body.isNotEmpty) {
-      /// path the user pin, with
+      /// parse the user from the response
       final IUser user = IUser.fromJson(json.decode(response.body));
-      await _patchPin(user.id!, flipperHttpClient, apihub,
-          ownerName: user.tenants.first.name!);
+
+      // Make PIN patching non-blocking for the login flow
+      try {
+        await _patchPin(user.id!, flipperHttpClient, apihub,
+            ownerName: user.tenants.first.name!);
+      } catch (e) {
+        // Log the error but don't block login
+        talker.warning("Failed to patch PIN, but continuing login: $e");
+        // This ensures offline login still works even if PIN patching fails
+      }
+
       ProxyService.box.writeInt(key: 'userId', value: user.id!);
 
       // Only perform Firebase login if not already logged in
@@ -327,6 +337,7 @@ mixin AuthMixin implements AuthInterface {
       required bool skipDefaultAppSetup,
       bool stopAfterConfigure = false,
       required Pin pin,
+      bool freshUser = false,
       required HttpClientInterface flipperHttpClient,
       IUser? existingUser}) async {
     final flipperWatch? w =
@@ -337,7 +348,8 @@ mixin AuthMixin implements AuthInterface {
     // Use existing user data if provided, otherwise make the API call
     print('Before _authenticateUser');
     final IUser user = existingUser ??
-        await _authenticateUser(phoneNumber, pin, flipperHttpClient);
+        await _authenticateUser(phoneNumber, pin, flipperHttpClient,
+            freshUser: freshUser);
     print('After _authenticateUser');
     await configureSystem(userPhone, user, offlineLogin: offlineLogin);
     print('After configureSystem');
@@ -379,8 +391,21 @@ mixin AuthMixin implements AuthInterface {
                 "Setting business preferences for ${selectedBusiness.name}");
             await ProxyService.box.writeString(
                 key: 'bhfId', value: (await ProxyService.box.bhfId()) ?? "00");
-            await ProxyService.box
-                .writeInt(key: 'tin', value: selectedBusiness.tinNumber ?? 0);
+
+            // Get existing tin value if available
+            final existingTin = ProxyService.box.readInt(key: 'tin');
+
+            // Only update tin if selectedBusiness.tinNumber is not null or there's no existing value
+            if (selectedBusiness.tinNumber != null || existingTin == null) {
+              await ProxyService.box.writeInt(
+                  key: 'tin',
+                  value: selectedBusiness.tinNumber ?? existingTin ?? 0);
+              talker.debug(
+                  'Setting tin to ${selectedBusiness.tinNumber ?? existingTin ?? 0} (from ${selectedBusiness.tinNumber != null ? 'business' : 'existing value'})');
+            } else {
+              talker.debug('Preserving existing tin value: $existingTin');
+            }
+
             await ProxyService.box.writeString(
                 key: 'encryptionKey',
                 value: selectedBusiness.encryptionKey ?? "");
@@ -824,5 +849,122 @@ mixin AuthMixin implements AuthInterface {
   }) async {
     // Add social login logic here
     return null;
+  }
+
+  /// Authenticates with Supabase using branch ID as email
+  @override
+  Future<void> supabaseAuth() async {
+    try {
+      // Get branch ID for email construction
+      final branchId = ProxyService.box.getBranchId();
+      if (branchId == null) {
+        talker.warning(
+            'Cannot authenticate with Supabase: No branch ID available');
+        return;
+      }
+
+      final email = '$branchId@flipper.rw';
+      talker.debug('Supabase email: $email');
+
+      // Check if we already have a valid session
+      final currentSession =
+          superUser.Supabase.instance.client.auth.currentSession;
+      final currentUser = superUser.Supabase.instance.client.auth.currentUser;
+
+      // Check if the session is still valid (not expired)
+      final bool hasValidSession = currentSession != null &&
+          currentUser != null &&
+          currentUser.email == email &&
+          _isSessionValid(currentSession);
+
+      if (hasValidSession) {
+        talker.debug(
+            'Supabase session is still valid, skipping re-authentication');
+        return;
+      }
+
+      talker.debug('No valid Supabase session found, authenticating...');
+
+      // Determine if we need to sign up or sign in
+      if (currentUser == null) {
+        // Try to sign up first
+        try {
+          final session =
+              superUser.Supabase.instance.client.auth.currentSession;
+          if (session != null) {
+            // User is already logged in.  Handle this situation.
+            talker.debug("User is already logged in.  Sign out first.");
+            await superUser.Supabase.instance.client.auth
+                .signOut(); // Sign out if necessary
+          }
+          superUser.AuthResponse auth =
+              await superUser.Supabase.instance.client.auth.signInWithPassword(
+            email: "1@flipper.rw",
+            password: "1@flipper.rw",
+          );
+
+          _saveSessionData(auth);
+          talker.debug('Supabase user created and signed in successfully');
+        } catch (signUpError) {
+          // If sign up fails (likely because user already exists), try sign in
+          talker.debug('Sign up failed, attempting sign in: $signUpError');
+          await _attemptSignIn(email);
+          await superUser.Supabase.instance.client.auth.signUp(
+            email: email,
+            password: email,
+          );
+        }
+      } else {
+        // User exists but session is invalid, just sign in
+        await _attemptSignIn(email);
+      }
+    } catch (e) {
+      talker.error('Supabase authentication error: $e');
+    }
+  }
+
+  /// Attempts to sign in with the given email
+  Future<void> _attemptSignIn(String email) async {
+    try {
+      superUser.AuthResponse auth =
+          await superUser.Supabase.instance.client.auth.signInWithPassword(
+        email: email,
+        password: email,
+      );
+
+      _saveSessionData(auth);
+      talker.debug('Supabase sign in successful');
+    } catch (signInError) {
+      talker.error('Supabase sign in failed: $signInError');
+      rethrow;
+    }
+  }
+
+  /// Saves session data to local storage
+  void _saveSessionData(superUser.AuthResponse auth) {
+    final expiresAt = auth.session?.expiresAt ?? 0;
+    final refreshToken = auth.session?.refreshToken ?? "";
+
+    ProxyService.box.writeString(key: 'refreshToken', value: refreshToken);
+    ProxyService.box.writeInt(key: 'expiresAt', value: expiresAt);
+
+    // Also store the access token for potential use elsewhere
+    final accessToken = auth.session?.accessToken ?? "";
+    ProxyService.box
+        .writeString(key: 'supabaseAccessToken', value: accessToken);
+  }
+
+  /// Checks if a session is still valid
+  bool _isSessionValid(superUser.Session session) {
+    // Get current time in seconds since epoch
+    final now = DateTime.now().millisecondsSinceEpoch ~/ 1000;
+
+    // Add a buffer of 5 minutes (300 seconds) to refresh before actual expiration
+    const expirationBuffer = 300;
+
+    // Session is valid if it's not expired (with buffer)
+    // Handle null expiresAt safely
+    final expiresAt = session.expiresAt ?? 0;
+    return expiresAt > (now + expirationBuffer);
   }
 }
