@@ -163,9 +163,10 @@ mixin TransactionItemMixin implements TransactionItemInterface {
     }
   }
 
+  @override
   Stream<List<TransactionItem>> transactionItemsStreams({
     String? transactionId,
-    int? branchId,
+    String? branchId,
     String? branchIdString,
     DateTime? startDate,
     DateTime? endDate,
