@@ -44,6 +44,8 @@ class CronService {
   /// - Periodically attempts to publish the device to the server
   Future<void> schedule() async {
     try {
+      ProxyService.strategy.businesses(
+          userId: ProxyService.box.getUserId() ?? 0, fetchOnline: true);
       await _initializeData();
       _setupPeriodicTasks();
       await _configureServices();
@@ -377,7 +379,8 @@ class CronService {
   /// Sets up Firebase messaging for notifications
   Future<void> _setupFirebaseMessaging() async {
     try {
-      Business? business = await ProxyService.strategy.getBusiness();
+      Business? business = await ProxyService.strategy
+          .getBusiness(businessId: ProxyService.box.getBusinessId()!);
       if (business == null) {
         talker.warning("Skipping Firebase messaging setup: Business is null");
         return;
